Added streaming.tut, a tutorial on cats.data.Streaming.
This tutorial mostly just shows some uses of Streaming[A].
It doesn't go as deeply as it could into the relationship
with Eval[A], or with things like Monad[A] and Traverse[A].

It also adds the Streaming syntax to implicits, and
renames concat to ++ for consistency.
