+++ core/src/main/scala/cats/syntax/package.scala
@@ -25,6 +25,7 @@ package object syntax {
   object semigroupk extends SemigroupKSyntax
   object show extends Show.ToShowOps
   object split extends SplitSyntax
+  object streaming extends StreamingSyntax
   object strong extends StrongSyntax
   object traverse extends TraverseSyntax
   object xor extends XorSyntax
