+++ tests/src/test/scala/cats/tests/StreamingTests.scala
@@ -6,6 +6,7 @@ import algebra.laws.OrderLaws
 import cats.data.Streaming
 import cats.laws.discipline.arbitrary._
 import cats.laws.discipline.{TraverseTests, CoflatMapTests, MonadCombineTests, SerializableTests}
+import cats.syntax.all._
 
 class StreamingTests extends CatsSuite {
   checkAll("Streaming[Int]", CoflatMapTests[Streaming].coflatMap[Int, Int, Int])
@@ -71,9 +72,9 @@ class AdHocStreamingTests extends CatsSuite {
     }
   }
 
-  test("concat") {
+  property("++") {
     forAll { (xs: List[Int], ys: List[Int]) =>
-      (convert(xs) concat convert(ys)).toList shouldBe (xs ::: ys)
+      (convert(xs) ++ convert(ys)).toList shouldBe (xs ::: ys)
     }
   }
 
@@ -177,7 +178,6 @@ class AdHocStreamingTests extends CatsSuite {
     }
   }
 
-  import Streaming.syntax._
   import scala.util.Try
 
   val bomb: Streaming[Int] =
@@ -216,12 +216,12 @@ class AdHocStreamingTests extends CatsSuite {
     bomb.peekEmpty shouldBe None
   }
 
-  test("lazy concat") {
-    isok(bomb concat bomb)
+  property("lazy ++") {
+    isok(bomb ++ bomb)
   }
 
-  test("lazier concat") {
-    isok(bomb concat Always(sys.error("ouch"): Streaming[Int]))
+  property("lazier ++") {
+    isok(bomb ++ Always(sys.error("ouch"): Streaming[Int]))
   }
 
   test("lazy zip") {
